+++ src/components/icons/clockOutlinedTeal.tsx
